+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.4.1 under development
 
-- New #377: Add `Color` field for HTML5 color input (@copilot)
+- New #377: Add `Color` field (@samdark)
 
 ## 1.4.0 March 27, 2025